update: learning_hub.py, mcp_servers.json, auto_optimizer.py +1 more

+++ tools/auto_optimizer.py
@@ -137,7 +137,7 @@ class AutoOptimizer:
             with conn.cursor() as cur:
                 cur.execute("""
                     SELECT agent_role,
-                           ROUND(AVG(latency_ms), 0) AS avg_latency,
+                           ROUND(AVG(latency_ms)::numeric, 0) AS avg_latency,
                            COUNT(*) AS total
                     FROM evaluations
                     WHERE latency_ms > 0
